1.0.3
Implement buffer feature, which uses an etsdb database to store updates that failed to send. Upon successfully sending an update, the DSLink will also send any unsent updates that are in the buffer. The buffer includes optional auto-purge functionality, which will erase the oldest unsent updates if the buffer exceeds a certain size. In order to more efficiently send the buffered updates, the rule format now includes %STARTBLOCK% and %ENDBLOCK% as delimeters for a repeatable block in a "batch update" request.